requirements.txt
playwright
openai
deepdiff
python-dotenv
beautifulsoup4